chore: refactored time_utils.py to utils/date.py
